+++ .travis.yml
@@ -10,8 +10,7 @@ services:
 before_install:
   - sudo apt-get -y install acl
 
-
-install: pip install -r ./test-requirements.txt
+install: pip install -r ./test-requirements.txt && pip install -r ./requirements.txt
 
 script: make check
