+++ conu/apidefs/__init__.py
@@ -1,4 +1,4 @@
 """
-This python module contains defintions in form of abstract classes which every
+This python module contains definitions in form of abstract classes which every
 backend needs to implement
 """
