+++ conu/apidefs/image.py
@@ -1,0 +1,10 @@
+"""
+Abstract definition for an Image
+"""
+
+
+class Image(object):
+    """
+    A class which represents an arbitrary container image. It contains utility methods
+    to manipulate it.
+    """
