+++ conu/backend/docker/container.py
@@ -1,80 +1,20 @@
-# -*- coding: utf-8 -*-
+"""
+Implementation of a docker container
+"""
 
 import json
 import subprocess
 
+from .image import DockerImage
+from conu.apidefs.container import Container
 from conu.utils.core import run_cmd, random_str, logger
 
 
-class Image(object):
-    """
-    A class which represents a docker container image. It contains utility methods to manipulate it.
-    """
-    def __init__(self, repository, tag=None):
-        """
-        :param repository: str, image name, examples: "fedora", "registry.fedoraproject.org/fedora",
-                            "tomastomecek/sen", "docker.io/tomastomecek/sen"
-        :param tag: str, tag of the image, when not specified, "latest" is implied
-        """
-        self.tag = tag or "latest"
-        self.name = repository
-        self.inspect_data = None
-        self.additional_names = []
-
-    def full_name(self):
-        return "%s:%s" % (self.name, self.tag)
-
-    def pull(self):
-        run_cmd("docker image pull %s" % self.full_name())
-
-    def tag_image(self, repository=None, tag=None):
-        """
-        Apply additional tags to the image or even add a new name
-
-        :param repository: str, see constructor
-        :param tag: str, see constructor
-        :return: str, the new name
-        """
-        if not (repository or tag):
-            raise ValueError("You need to specify either repository or tag.")
-        t = repository or self.name
-        t = "%s:%s" % (t, "latest" if not tag else tag)
-        run_cmd("docker image tag %s %s" % (self.full_name(), t))
-        self.additional_names.append(t)
-        return t
-
-    def __repr__(self):
-        return "Image(repository=%s, tag=%s)" % (self.name, self.tag)
-
-    def __str__(self):
-        return self.full_name()
-
-    def inspect(self, force=False):
-        if force or not self.inspect_data:
-            self.inspect_data = json.loads(run_cmd("docker image inspect %s" % self.full_name()))[0]
-        return self.inspect_data
-
-    @classmethod
-    def rmi(cls, image, force=False):
-        """
-        remove selected image
-
-        :param image: str, image name, example: "fedora:latest"
-        :return: None
-        """
-        run_cmd("docker image remove %s%s" % ("-f " if force else "",  image))
-
-    def clean(self, force=False):
-        images_to_remove = self.additional_names + [self.full_name()]
-        for i in images_to_remove:
-            Image.rmi(i, force=force)
-        self.additional_names = []
-
-
-class Container(object):
+class DockerContainer(Container):
     def __init__(self, image, tag=None):
+        super(DockerContainer, self).__init__(image, None)
         self.tag = tag or random_str()
-        if not isinstance(image, Image):
+        if not isinstance(image, DockerImage):
             raise BaseException("Image is not instance of Image class")
         self.json = None
         self.image = image
@@ -155,7 +95,8 @@ class Container(object):
     def run(self, command="", docker_params="", **kwargs):
         command = command.split(" ") if command else []
         additional_params = docker_params.split(" ") if docker_params else []
-        cmdline = ["docker", "container", "run", "--name", self.tag] + additional_params + [self.image.full_name()] + command
+        cmdline = ["docker", "container", "run", "--name", self.tag] + additional_params + \
+                  [self.image.full_name()] + command
         output = run_cmd(cmdline, **kwargs)
         if not self.__occupied:
             self.clean()
