--- conu/docker/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from core import *
