+++ conu/utils/__init__.py
@@ -1,3 +1,1 @@
 # -*- coding: utf-8 -*-
-
-from core import *
